+++ app/components/HeaderComponent/HeaderComponent.tsx
@@ -1,0 +1,5 @@
+const HeaderComponent = ()=> {
+    return <></>
+}
+
+export default HeaderComponent
